+++ src/bentopy/pack/space.py
@@ -177,7 +177,7 @@ class Space:
 
         # Apply an offset to the spots that are considered valid based on the possible center adjustment for a segment.
         # relevant = np.array([1.0 if v is None else 0.0 for v in center])
-        center_offset = segment.center_translation() # FIXME dtype -> int
+        center_offset = segment.center_translation(dtype=int)
         for compartment in self.compartments:
             if compartment.id not in segment.compartment_ids:
                 continue
